backend/auth-service/requirements.txt
fastapi
uvicorn
pydantic>=2.7.0
python-dotenv
google-cloud-firestore>=2.14.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
firebase-admin>=6.5.0
